refactor: move modals to components folder
